Private Chat Room

+++ CommunityConnection.Common/MessageResponse.cs
@@ -8,7 +8,7 @@ namespace CommunityConnection.Common
 {
     public class MessageChannelResponse
     {
-        public int ChannelId { get; set; }
+        public long ChannelId { get; set; }
         public List<MessageResponse> Messages { get; set; }
     }
     public class MessageDetail
