+++ CommunityConnection.Infrastructure/Repository/IMessageChannelRepository.cs
@@ -1,4 +1,5 @@
 ﻿using CommunityConnection.Common;
+using CommunityConnection.Entities.DTO;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,8 +10,10 @@ namespace CommunityConnection.Infrastructure.Repository
 {
     public interface IMessageChannelRepository
     { 
-        Task<List<MessageResponse>> GetMessagesAsync(long userId, int channelId);
+        Task<List<MessageResponse>> GetMessagesAsync(long userId, long channelId);
         Task<bool> IsChannelMemberAsync(long channelId, long userId);
         Task<Message> CreateMessageAsync(Message message);
+
+
     }
 }
